refactor new blocks
