+++ express/blocks/link-list/link-list.js
@@ -2,8 +2,44 @@ import { getLibs } from '../../scripts/utils.js';
 import buildCarousel from '../shared/carousel.js';
 import { decorateButtons_deprecated, addTempWrapper } from '../../scripts/utils/decorate.js';
 
-let { getConfig } = await import(`${getLibs()}/utils/utils.js`);
+const { getConfig } = await import(`${getLibs()}/utils/utils.js`);
 
+export function getHelixEnv() {
+  let envName = sessionStorage.getItem('helix-env');
+  if (!envName) {
+    envName = 'stage';
+    if (window.spark?.hostname === 'www.adobe.com') envName = 'prod';
+  }
+  const envs = {
+    stage: {
+      commerce: 'commerce-stg.adobe.com',
+      adminconsole: 'stage.adminconsole.adobe.com',
+      spark: 'stage.projectx.corp.adobe.com',
+    },
+    prod: {
+      commerce: 'commerce.adobe.com',
+      spark: 'express.adobe.com',
+      adminconsole: 'adminconsole.adobe.com',
+    },
+  };
+  const env = envs[envName];
+
+  const overrideItem = sessionStorage.getItem('helix-env-overrides');
+  if (overrideItem) {
+    const overrides = JSON.parse(overrideItem);
+    const keys = Object.keys(overrides);
+    env.overrides = keys;
+
+    for (const a of keys) {
+      env[a] = overrides[a];
+    }
+  }
+
+  if (env) {
+    env.name = envName;
+  }
+  return env;
+}
 
 async function fetchRelevantRows(path) {
   if (!window.relevantRows) {
@@ -119,7 +155,6 @@ export default async function decorate(block) {
 
     const platformEl = document.createElement('div');
     platformEl.classList.add('link-list-platform');
-    console.log("build Carousel");
     await buildCarousel('p.button-container', block, options);
   }
