+++ express/blocks/steps/steps.js
@@ -1,5 +1,6 @@
 import { getLibs } from '../../scripts/utils.js';
-const { createTag, getConfig } = await import(`${getLibs()}/utils/utils.js`);
+
+const { createTag } = await import(`${getLibs()}/utils/utils.js`);
 
 export function addBlockClasses($block, classNames) {
   const $rows = Array.from($block.children);
